Split out classes into individual files Should make project a bit easier to manage, imports added to __init__.py for backwards compatibility
